+++ src/replacements/program_options/errors.h
@@ -1,3 +1,8 @@
+// Copyright Vladimir Prus 2004, modified by Dmitry Kargin 2024.
+// Distributed under the Boost Software License, Version 1.0.
+// (See accompanying file LICENSE_1_0.txt
+// or copy at http://www.boost.org/LICENSE_1_0.txt
+
 #ifndef MINIROS_PROGRAM_OPTIONS_ERRORS_H
 #define MINIROS_PROGRAM_OPTIONS_ERRORS_H
