+++ src/replacements/program_options/program_options.h
@@ -1,6 +1,7 @@
-/// A local replacement of boost::program_options.
-/// It should support usage scenarios of miniros/rosbag.
-/// No other features are needed.
+// Copyright Vladimir Prus 2004, modified by Dmitry Kargin 2024.
+// Distributed under the Boost Software License, Version 1.0.
+// (See accompanying file LICENSE_1_0.txt
+// or copy at http://www.boost.org/LICENSE_1_0.txt
 
 #ifndef MINIROS_PROGRAM_OPTIONS_H
 #define MINIROS_PROGRAM_OPTIONS_H
