+++ src/replacements/program_options/values.inl
@@ -1,6 +1,12 @@
+// Copyright Vladimir Prus 2004, modified by Dmitry Kargin 2024.
+// Distributed under the Boost Software License, Version 1.0.
+// (See accompanying file LICENSE_1_0.txt
+// or copy at http://www.boost.org/LICENSE_1_0.txt
+
 #ifndef MINIROS_PROGRAM_OPTIONS_VALIDATORS_INL
 #define MINIROS_PROGRAM_OPTIONS_VALIDATORS_INL
 
+#include <cassert>
 #include <sstream>
 #include <any>
 #include <string>
