+++ src/replacements/program_options/variables_map.cpp
@@ -1,13 +1,109 @@
+// Copyright Vladimir Prus 2004, modified by Dmitry Kargin 2024.
+// Distributed under the Boost Software License, Version 1.0.
+// (See accompanying file LICENSE_1_0.txt
+// or copy at http://www.boost.org/LICENSE_1_0.txt
+
 #include <cassert>
 #include <map>
 #include <set>
 
-#include "replacements/program_options/variables_map.h"
+#include "variables_map.h"
 
-#include "replacements/program_options/option.h"
-#include "replacements/program_options/program_options.h"
+#include "option.h"
+#include "program_options.h"
 
 namespace program_options {
+
+
+abstract_variables_map::abstract_variables_map()
+: m_next(0)
+{}
+
+abstract_variables_map::
+abstract_variables_map(const abstract_variables_map* next)
+: m_next(next)
+{}
+
+const variable_value&
+abstract_variables_map::operator[](const std::string& name) const
+{
+    const variable_value& v = get(name);
+    if (v.empty() && m_next)
+        return (*m_next)[name];
+    else if (v.defaulted() && m_next) {
+        const variable_value& v2 = (*m_next)[name];
+        if (!v2.empty() && !v2.defaulted())
+            return v2;
+        else return v;
+    } else {
+        return v;
+    }
+}
+
+void abstract_variables_map::next(abstract_variables_map* next)
+{
+    m_next = next;
+}
+
+///////////////////////////////////////////////////////////////////////////////////
+variables_map::variables_map()
+{}
+
+variables_map::variables_map(const abstract_variables_map* next)
+: abstract_variables_map(next)
+{}
+
+void variables_map::clear()
+{
+    std::map<std::string, variable_value>::clear();
+    m_final.clear();
+    m_required.clear();
+}
+
+const variable_value& variables_map::get(const std::string& name) const
+{
+    static variable_value empty;
+    const_iterator i = this->find(name);
+    if (i == this->end())
+        return empty;
+    else
+        return i->second;
+}
+
+void variables_map::notify()
+{
+    // This checks if all required options occur
+    for (map<std::string, std::string>::const_iterator r = m_required.begin();
+         r != m_required.end();
+         ++r)
+    {
+        const std::string& opt = r->first;
+        const std::string& display_opt = r->second;
+        std::map<std::string, variable_value>::const_iterator iter = find(opt);
+        if (iter == end() || iter->second.empty())
+        {
+            throw required_option(display_opt);
+        }
+    }
+
+#ifdef SUPPORT_NOTIFY
+    // Lastly, run notify actions.
+    for (std::map<std::string, variable_value>::iterator k = begin();
+         k != end(); ++k)
+    {
+        /* Users might wish to use variables_map to store their own values
+           that are not parsed, and therefore will not have value_semantics
+           defined. Do not crash on such values. In multi-module programs,
+           one module might add custom values, and the 'notify' function
+           will be called after that, so we check that value_sematics is
+           not NULL. See:
+               https://svn.boost.org/trac/boost/ticket/2782
+        */
+        if (k->second.m_value_semantic)
+            k->second.m_value_semantic->notify(k->second.value());
+    }
+#endif
+}
 
 // First, performs semantic actions for 'oa'.
 // Then, stores in 'm' all options that are defined in 'desc'.
